+++ packages/gamut/src/Box/props.ts
@@ -14,8 +14,46 @@ export const boxProps = variance.compose(
   system.grid
 );
 
-export interface BoxProps extends StyleProps<typeof boxProps> {}
+export const sharedStates = system.states({
+  fit: {
+    width: 1,
+    height: 1,
+  },
+});
 
-export interface FlexBoxProps extends BoxProps {}
+export const flexStates = system.states({
+  inline: {
+    display: 'inline-flex',
+  },
+  wrap: {
+    flexWrap: 'wrap',
+  },
+  center: {
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  row: {
+    flexDirection: 'row',
+  },
+  column: {
+    flexDirection: 'column',
+  },
+});
 
-export interface GridBoxProps extends BoxProps {}
+export const gridStates = system.states({
+  center: {
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  fitContent: {
+    gridTemplate: 'minmax(0, 1fr) / minmax(0, 1fr)',
+  },
+});
+
+export interface BoxProps
+  extends StyleProps<typeof boxProps>,
+    StyleProps<typeof sharedStates> {}
+
+export interface FlexBoxProps extends BoxProps, StyleProps<typeof flexStates> {}
+
+export interface GridBoxProps extends BoxProps, StyleProps<typeof gridStates> {}