+++ packages/gamut-styles/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [11.2.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@11.1.0...@codecademy/gamut-styles@11.2.0) (2021-06-10)
+
+
+### Features
+
+* Navy scale + color alias changes GM-237 ([bf896dd](https://github.com/Codecademy/client-modules/commit/bf896dd3fcf0435ef557694740d96df38d12173b)), closes [#10162](https://github.com/Codecademy/client-modules/issues/10162)
+
+
+
+## [11.1.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@11.0.0...@codecademy/gamut-styles@11.1.0) (2021-06-10)
+
+
+### Features
+
+* **variance:** states boolean toggle styles GM-233 ([e463f0e](https://github.com/Codecademy/client-modules/commit/e463f0e77802a3565ddda4de2e14050372923678))
+
+
+
 ## [11.0.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-styles@10.2.1...@codecademy/gamut-styles@11.0.0) (2021-06-03)